+++ src/recon_tmpl.c
@@ -1225,7 +1225,7 @@ static int mc(Dav1dTaskContext *const t,
         const int orig_pos_x = (bx * h_mul << 4) + mvx * (1 << !ss_hor);
 #define scale_mv(res, val, scale) do { \
             const int64_t tmp = (int64_t)(val) * scale + (scale - 0x4000) * 8; \
-            res = apply_sign64((int) ((llabs(tmp) + 128) >> 8), tmp) + 32;     \
+            res = apply_sign64((llabs(tmp) + 128) >> 8, tmp) + 32;     \
         } while (0)
         int pos_y, pos_x;
         scale_mv(pos_x, orig_pos_x, f->svc[refidx][0].scale);
